backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
deepface
opencv-python-headless
tf-keras
python-multipart
Pillow
pillow-heif
aiofiles
numpy<2
ruff
mypy
types-requests
